introduce new data struct, for check that the data is correct, change int to uchar for color struct

+++ src/touch/datastruct/datastruct.h
@@ -15,7 +15,7 @@
 #define NCOLOR 4
 
 struct colore_s{
-    int colore[NCOLOR];
+    uchar colore[NCOLOR];
 };
 
 struct point_s{
